requirements.txt
fastapi==0.115.2
strawberry-graphql==0.243.1
langgraph==0.2.34
pydantic==2.9.2
uvicorn==0.32.0
sqlalchemy==2.0.36
pymongo==4.10.1
redis==5.1.1
pyjwt==2.9.0
pytest==8.3.3
httpx==0.27.2
prometheus-client==0.21.0
python-dotenv==1.0.1
